requirements.txt
SpeechRecognition
PyAudio
pyautogui
pyttsx3
requests
keyboard
python-dotenv
